Payment, Role and Ticket systems.

+++ src/main/java/com/devtiro/quickstart/repository/RolesRepository.java
@@ -1,8 +1,8 @@
 package com.devtiro.quickstart.repository;
 
-import com.devtiro.quickstart.model.Roles;
+import com.devtiro.quickstart.entity.Roles;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface RolesRepository extends JpaRepository<Roles, Integer> {
+public interface RolesRepository extends JpaRepository<Roles, Long> {
 }
